requirements.txt
streamlit
streamlit_drawable_canvas
requests
pillow
torch
torchvision
